requirements.txt
streamlit==1.22.0
tensorflow==2.20.0
numpy==1.24.3
Pillow==9.5.0
